requirements.txt
altgraph==0.17.3
artifactory==0.1.17
begins==0.9
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
Cython==3.0.2
easycython==1.0.7
fonttools==4.42.1
gevent==22.10.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
isort==5.13.2
Jinja2==3.1.2
kiwisolver==1.4.5
loguru==0.7.0
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mdurl==0.1.2
msl-loadlib==0.9.0
mypy==1.9.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
numpy==1.24.4
packaging==23.1
pathspec==0.11.1
pefile==2023.2.7
Pillow==10.0.1
platformdirs==3.5.0
protobuf==3.20.3
psutil==5.9.5
pyarmor==8.3.9
pyarmor.cli.core==4.3.5
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.8
pyparsing==3.1.1
pypiwin32==223
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
requests==2.31.0
rich==13.3.5
robotframework==6.0.2
robotframework-gevent==0.5.1
robotframework-pythonlibcore==4.1.2
rtcclient==0.6.0
schedule==1.2.0
shellingham==1.5.0.post1
six==1.16.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
tomli==2.0.1
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.5.0
urllib3==2.0.3
win32-setctime==1.1.0
xmltodict==0.13.0
zipp==3.17.0
zope.event==4.6
zope.interface==6.0
